todo-bot/requirements.txt
fastapi
openai
uvicorn
pydantic
python-dotenv
requests
bubbletea-chat[llm]
